move types to DB
