+++ src/jobs/updated_ats_until.rs
@@ -1,8 +1,8 @@
 use chrono::NaiveDateTime;
 
-use crate::database::{self, RequestBuilder};
+use crate::database::{self, DBResultTypes, RequestBuilder};
 use crate::diff::IO;
-use crate::{Config, DBResultTypes};
+use crate::Config;
 
 use super::par_run;
